requirements.txt
# OCR & 이미지 처리
ultralytics
paddlepaddle==3.2.2 -i https://www.paddlepaddle.org.cn/packages/stable/cpu/\
paddleocr
opencv-python
Pillow
numpy
#accelerate

# LLM & ML
torch
transformers
sentence-transformers

# VectorDB
chromadb

# API & 검색
openai
python-dotenv
google-api-python-client
youtube_transcript_api

# UI
streamlit
